Reorganized so I could import common test functions.
