+++ backend/tests/pytests/integration/rest_api/common.py
@@ -1,0 +1,37 @@
+from uuid import uuid1
+
+from offle_assistant.main import create_default_admin
+
+
+async def create_test_user(client) -> str:
+    uuid_str = uuid1()
+    user_email = f"{uuid_str}@example.com"
+    password = "securepassword"
+
+    reg_user_payload = {
+        "email": user_email,
+        "password": password,
+    }
+
+    reg_user_response = await client.post(
+        "/auth/register", json=reg_user_payload)
+    data = reg_user_response.json()
+    return {
+        "user_id": data["user_id"],
+        "email": user_email,
+        "password": password
+    }
+
+
+async def get_default_admin_token(client) -> str:
+    await create_default_admin()
+    default_admin_payload = {
+        "email": "admin@admin.com",
+        "password": "admin",
+    }
+    login_response = await client.post(
+        "/auth/login", json=default_admin_payload
+    )
+
+    data = login_response.json()
+    return data["access_token"]
